back/requirements.txt
asyncio
pyserial
DHT11
